+++ examples/wpf-contacts/WpfContacts/App.xaml.cs
@@ -1,0 +1,7 @@
+using System.Windows;
+
+namespace WpfContacts;
+
+public partial class App : Application
+{
+}